+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 GoOracle Changes
 ----------------
-
+## 0.2
+    * Complete rewrite
+    * Added commands from quick menu
+    * 
 ## 0.1.1
     * Optionally place your cursor over a symbol instead of selecting it.